Separate the code related to RollingUpdate
